Fix interactive password reading
restic no longer tries to read passwords from stdin. Instead, it tries
to open the controlling terminal and fails when that is not present.

This changes some surprising behavior when reading passwords
interactively, most notably in restic backup --stdin:

	$ echo 'Hello, world!' | restic -r t backup --stdin
	read password from stdin
	Fatal: wrong password or no key found

Absence of a controlling terminal would lead to even stranger behavior:

	$ setsid restic -r t snapshots
	$ enter password for repository:
	tFatal: wrong password or no key found

The 't' in tFatal is part of the password echoing.

As a side-effect, some of the code has been streamlined. The former
ReadPasswordTwice would read the password once when stdin was not a
terminal, breaking its own contract.